+++ flutter_searchbox_ui/CHANGELOG.md
@@ -3,7 +3,7 @@
 - Allows custom stlyes to be set.
 - Fixes broken ui when clearing dropdown.
 - Triggers query with default values.
-- Supports unbounded upper limit via `no limit` dropdown option.
+- Supports unbounded upper limit via `no_limit` dropdown option.
 ## [1.0.2-alpha] - 30-09-2021.
 
 - Simplified RangeInput Example.